Move functions generateUrl() and parseUrl()
