+++ src/ts/components/FormikMain.tsx
@@ -1,9 +1,8 @@
 import * as SemanticUiReact from 'semantic-ui-react';
 import * as React from 'react';
 import FormikGrid from './FormikGrid';
-import { ObjectTable } from '../logic/table-data';
+import { arrayTableToObjectTable,ObjectTable,objectTableToArrayTable } from '../logic/table-data';
 import { ColumnsDefinition } from '../logic/query-binder';
-import { parseUrl, generateUrl } from '../containers/Operator';
 import * as UrlBinder from '../logic/url-binder';
 
 interface Stringified {
@@ -143,4 +142,34 @@ export default class FormikMain extends React.Component<Props, State> {
             structured: this.state.structured
         });
     }
+}
+
+interface ObjectTables {
+    basic: ObjectTable;
+    coord: ObjectTable;
+    host: ObjectTable;
+    libs: ObjectTable;
+}
+
+function generateUrl(tables: ObjectTables): string {
+    return UrlBinder.generateUrl({
+        host:
+            objectTableToArrayTable(UrlBinder.ColumnsDefinition.host, tables.host),
+        query: {
+            basic: objectTableToArrayTable(ColumnsDefinition.basic, tables.basic),
+            coord: objectTableToArrayTable(ColumnsDefinition.coord, tables.coord),
+            libs: objectTableToArrayTable(ColumnsDefinition.libs, tables.libs)
+        }
+    });
+}
+
+function parseUrl(url: string): ObjectTables {
+    const parsed = UrlBinder.parseUrl(url);
+
+    return {
+        basic: arrayTableToObjectTable(ColumnsDefinition.basic, parsed.query.basic),
+        coord: arrayTableToObjectTable(ColumnsDefinition.coord, parsed.query.coord),
+        host: arrayTableToObjectTable(UrlBinder.ColumnsDefinition.host, parsed.host),
+        libs: arrayTableToObjectTable(ColumnsDefinition.libs, parsed.query.libs)
+    };
 }
